refactor: 폴더 구조 변경
- components 폴더 생성
  - Car, RacingCountForm

- validators 폴더 생성
  - CarNameValidator, TryCountValidator
